Fix base/target pair for Nova
The base/target pairs coming the their API endpoints are buggy. For
e.g., consider https://novaexchange.com/remote/v2/market/info/BTC_LTC/.

As of now, the base currency is BTC, target currency is LTC, but the
price is ~ 0.0189. This value is closer to the rate for LTC/BTC, not
BTC/LTC. Hence, the correct base should be LTC, and target should be
BTC. As a result, the market name should be target_base.